Return actual boolean from exist functions
The data-access API has certain "exist" functions, which are intended to
return a boolean that states whether a certain object exists or not.

The current SQL implementation ditches the boolean and, instead, throws
an ObjectNotFoundException if the object does not exist. While this
works at the end of the day, it's really an ugly hack and therefore
hardly reference material. This commit intends to fix the status quo.

While these changes should in theory be reflected in the rdap-server
project (because, as it is, the server is currently also ditching the
boolean), this commit will not have a rdap-server counterpart. The
reason is that I just discovered that we're implementing HEADs in a
non-standard way, and I will end up having to refactor the "exists"
funtions out of the API anyway. This essentially means that this commit
is just a checkpoint in case we ever decide to return those functions.

I will add more information regarding the HEAD bug in the next commit.