Delete HEAD-specific API/functions
According to the HTTP spec,

	The server SHOULD send the same header fields in response to a
	HEAD request as it would have sent if the request had been a GET
	(RFC 7231)

This essentially trumps the optimizations we had set up for HEAD
requests. The problem is the "content-length" header field; if we're
skipping the full queries in favor of a lightweight boolean one, there
is no way to guess the response size a GET would have had.

So instead of handling HEADs differently, we now fall back to
HttpServlet's default implementation of doHead().

On the plus side, the data-access API just became significantly more
straightforward to implement.