Adds dynamic-schema to the sql-provider
This commit enables the use of another schema in the database, 
different from the default value 'rdap'.

Add a configuration table to the database.

It creates a timer for the server, this timer has to be enabled from 
the configuration file, this will allow to switch between schemas when 
you want to change to a schema with more current data for the server 
rdap, making use of the configuration table to Know which is the most 
current scheme.