Rename main package
The current main package (the one that parents everything) of the
SQL classes is "mx.nic.rdap.db".
The current main package of the Data Access API classes is also
"mx.nic.rdap.db".

Though semantically incorrect, this wouldn't be a problem normally.

However, I need to review rdap-server doesn't directly reference
any SQL classes, and this is difficult if the package is not
indicative of the project the class belongs to. Hence this fix.