Try harder to find the data source
Sometimes, the JDNI name is "java:/comp/env/jdbc/rdap".
Other times, it's "java:comp/env/jdbc/rdap".

JBoss/Wildfly seems to consistently need the slash,
Glassfish/Payara seems to consistently reject it, and Tomcat seems
to not care. But I'm not sure. I tried reading their sources, but
I'm running out of patience and time. Oracle's documentation is
inconsistent as well; some examples include it, others do not.

In the previous commit I made the JDNI name configurable, which is
fine, but when this configuration is not present, since the slash
is the only source of problems, the provider will now attempt to
find the data source via both names before giving up.