Allow the user to write their own SQL files
The files must follow a key-value format.

Files should be added to the "classes / META-INF" folder on the server
by creating a folder called "user_sql_files" or in the properties of the
implementation by typing the "sql_files_directory" property and
indicating the path from the root of the disk to the folder where are
the SQL files.